[Test] Setup test and add test for bday cmd (#10)
The birthday commands are the most important part
and the parsing is regex. Added tests for a handful
of common cases and a few common incorrect cases to
make sure results are expected.

Did a bunch of quick refactoring for cleanup to make
test easier, and added a test job, coverage files etc.